# incubator-backend/openapi.yaml
openapi: 3.0.3
info:
  title: IT Youth Talent Incubator API
  description: |
    # IT Youth Talent Incubator Backend API
    
    Welcome to the comprehensive API documentation for the IT Youth Talent Incubator platform.
    This platform connects talented IT students in Ghana with companies seeking skilled professionals.
    
    ## Features
    - **Authentication & Authorization**: Secure JWT-based authentication with role-based access control
    - **User Management**: Support for Students, Companies, and Admins with distinct profiles
    - **Job Posting & Management**: Companies can post, manage, and track jobs
    - **Application System**: Students can apply to jobs and track application status
    - **Admin Panel**: Comprehensive admin tools for user approval and platform management
    
    ## Authentication
    Most endpoints require authentication via JWT tokens. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your_token>
    ```
    
    ## Rate Limiting
    API requests are rate-limited to prevent abuse:
    - General endpoints: 100 requests per 15 minutes
    - Authentication endpoints: 3-5 requests per 15 minutes
    
    ## Pagination
    List endpoints support pagination with the following query parameters:
    - `page`: Page number (default: 1)
    - `limit`: Items per page (default: 10, max: 100)
    
  version: 1.0.0
  contact:
    name: IT Youth Talent Incubator
    email: support@itfy-incubator.com
  license:
...
servers:
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login or registration
  schemas:
...
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "An error occurred"
        error:
          type: string
          example: "Detailed error message"
        code:
          type: string
          example: "ERR_CODE"
...
  responses:
    UnauthorizedError:
      description: Authentication token is missing or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            message: "Unauthorized"
            error: "You are not authorized to perform this action"
    ForbiddenError:
      description: User does not have permission to access this resource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            message: "Forbidden"
            error: "Access denied. You don't have permission for this resource"
...
paths:
...
  /jobs/bulk-update:
    post:
      tags:
        - Jobs
      summary: Bulk update job status (Company only)
      description: Update status for multiple jobs at once
      operationId: bulkUpdateStatus
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                jobIds:
                  type: array
                  items:
                    type: string
                status:
                  type: string
                  enum: [draft, published, paused, closed]
      responses:
        '200':
          description: Jobs updated successfully
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
...